DROP DOWN DEVELOPMENT style.css:I created the css variable to represent the base delay and duration for the drop down animation. I then applied those variables to the transition property of the drop down menu. This allows me to easily adjust the speed and delay of the animation by simply changing the values of the variables. This makes it easier to maintain and update the animation in the future.
added the calculation for the transition property of the segmentedDropDown
class so the duration will be set to the duration modifier times the base duration,
and the delay will be set to the delay modifier times the base delay.

and I commented out the keyframes for the flapDown animation
since I am no longer using it.